pyproject.toml
[project]
name = "taskor"
version = "1.4.5"
description = " A command-line interface tool for task automation and prompt management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42.0",
    "bs4>=0.0.2",
    "docx>=0.2.4",
    "ebooklib>=0.18",
    "markdown2>=2.5.2",
    "openai>=1.58.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pypdf2>=3.0.1",
    "pyperclip>=1.9.0",
    "pyqt5>=5.15.11",
    "pytesseract>=0.3.13",
    "pytest>=8.3.4",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "striprtf>=0.0.28",
    "xlrd>=2.0.1",
]

[project.scripts]
taskor = "taskor.taskor:main"

[tool.setuptools]
packages = ["taskor", "taskor.utils", "taskor.resources"]

[tool.setuptools.package-data]
taskor = ["resources/*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
